feat: 회원 정보 조회

+++ src/main/java/com/cogging/cogging/repository/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findById(Long id);
+    Optional<Member> findById(int id);
     Optional<Member> findByEmail(String email);
     Optional<Member> findByNickname(String nickname);
 }
